+++ scripts/const.js
@@ -1,0 +1,18 @@
+export const SPELLS = {
+  SUMMON: {
+    SUMMON_DRAGON: "Compendium.pf2e.spells-srd.Item.kghwmH3tQjMIhdH1",
+    SUMMON_UNDEAD: "Compendium.pf2e.spells-srd.Item.9WGeBwIIbbUuWKq0",
+    SUMMON_CELESTIAL: "Compendium.pf2e.spells-srd.Item.lTDixrrNKaCvLKwX",
+    SUMMON_FEY: "Compendium.pf2e.spells-srd.Item.hs7h8f4Z1ZNdUt3s",
+    SUMMON_ANIMAL: "Compendium.pf2e.spells-srd.Item.4YnON9JHYqtLzccu",
+    SUMMON_CONSTRUCT: "Compendium.pf2e.spells-srd.Item.lKcsmeOrgHtK4xQa",
+    SUMMON_LESSER_SERVITOR: "Compendium.pf2e.spells-srd.Item.B0FZLkoHsiRgw7gv",
+    SUMMON_PLANT_OR_FUNGUS: "Compendium.pf2e.spells-srd.Item.jSRAyd57kd4WZ4yE",
+    SUMMON_ELEMENTAL: "Compendium.pf2e.spells-srd.Item.lpT6LotUaQPfinjj",
+    SUMMON_ENTITY: "Compendium.pf2e.spells-srd.Item.i1TvBID5QLyXrUCa",
+    SUMMON_FIEND: "Compendium.pf2e.spells-srd.Item.29ytKctjg7qSW2ff",
+    SUMMON_GIANT: "Compendium.pf2e.spells-srd.Item.e9UJoVYUd5kJWUpi",
+    SUMMON_MONITOR: "Compendium.pf2e.spells-srd.Item.ZbEHglw5tkJ3grQZ",
+  },
+  INCARNATE: {},
+};