+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-## 
+## 1.0.0
+
+- `New`
+  - Includes support for all official spells with the `Summon` trait
+  - When a Spell with the summon tag is `Cast` it will pop up with the Foundry Summons Interface with the appropriate list of summons